main tests
