+++ main.cpp
@@ -3,17 +3,51 @@
 //
 
 #include <iostream>
+#include <fstream>
 #include "kraken.h"
+#include "compress.h"
+
+void error(const char *s, const char *curfile = NULL) {
+    if (curfile)
+        fprintf(stderr, "%s: ", curfile);
+    fprintf(stderr, "%s\n", s);
+    exit(1);
+}
+
+byte *load_file(const char *filename, int *size) {
+    FILE *f = fopen(filename, "rb");
+    if (!f) error("file open error", filename);
+    fseek(f, 0, SEEK_END);
+    int packed_size = ftell(f);
+    fseek(f, 0, SEEK_SET);
+    byte *input = new byte[packed_size];
+    if (!input) error("memory error", filename);
+    if (fread(input, 1, packed_size, f) != packed_size) error("error reading", filename);
+    fclose(f);
+    *size = packed_size;
+    return input;
+}
 
 int main() {
-    std::cout << "hello world" << std::endl;
+    const char *curfile = "E:\\GitHub\\kraken\\oodle.txt";
 
+    int input_size;
+    byte *input = load_file(curfile, &input_size);
 
+    byte *output = NULL;
+    output = new byte[input_size + 65536];
 
-    size_t src_len = 255;
-    size_t dst_len = 255;
+    int outbytes1 = Kraken_Compress( input, input_size, output);
+    int outbytes = CompressBlock(8, input, output + 8, input_size, 2, 0, 0, 0);
 
-    //Kraken_Decompress(nullptr, src_len, nullptr, dst_len);
+    std::ofstream file("E:\\GitHub\\kraken\\oodle.kark", std::ios::binary);
+
+    if (file)
+    {
+        file.write((char*) output, sizeof(output));
+    }
+
+    file.close();
 
     return 0;
 }
